+++ cli/src/cast.rs
@@ -21,10 +21,6 @@ use ethers::{
     types::{Address, NameOrAddress, U256},
 };
 
-// ZACH ADDITIONS
-use std::collections::HashMap;
-// END
-
 use eyre::WrapErr;
 use foundry_common::fs;
 use foundry_config::Chain;
@@ -215,187 +211,7 @@ async fn main() -> eyre::Result<()> {
                         "No Etherscan API Key is set. Consider using the ETHERSCAN_API_KEY env var, or setting the -e CLI argument or etherscan-api-key in foundry.toml"
                 ))?
             });
-
-            if let Some(hash) = tx_hash {
-
-                // Get the receipt from the tx hash and pull the logs.
-                let receipt = Cast::new(&provider).receipt(hash, None, 0, false, true).await?;
-                let receipt: serde_json::Value = serde_json::from_str(&receipt)?;
-                let logs = &receipt["logs"];
-                let mut event_abis: HashMap<String, Vec<String>> = HashMap::new();
-
-                let mut event_return_data: Vec<EventData> = Vec::new();
-
-                // For each event in the logs...
-                if let serde_json::Value::Array(logs) = logs {
-                    for event in logs {
-                        let address: &str = &event["address"].as_str().ok_or(
-                            eyre::eyre!("Could not parse event address from receipt"))?;
-                        let topics = &event["topics"].as_array().ok_or(
-                            eyre::eyre!("Could not parse topic log from receipt"))?;
-                        let topic0 =topics[0].as_str().ok_or(
-                            eyre::eyre!("Could not parse Topic 0 from receipt"))?;
-                        let mut data: String = event["data"].as_str().ok_or(
-                            eyre::eyre!("Could not parse data from receipt"))?.to_string();
-                        
-                        let mut event_data = EventData {
-                            sig: None,
-                            address: address.to_string(),
-                            args: Vec::new(),
-                        };
-                        
-                        let abi: Vec<String>;
-                        let mut args: Vec<EventArg> = Vec::new();
-                        
-                        // First, check whether the we already have the cached event abi from a previous event.
-                        if let Some(cached_abi) = event_abis.get(address) {
-                            abi = cached_abi.clone();
-                        } 
-                        // Otherwise, try to get the interface from etherscan & save to cache.
-                        else {
-                            let interfaces = SimpleCast::generate_interface(InterfacePath::Etherscan {
-                                chain: chain.inner,
-                                api_key: api_key.clone(),
-                                address: address
-                                    .parse::<Address>()
-                                    .wrap_err("Invalid address provided. Did you make a typo?")?,
-                            })
-                            .await?;
-            
-                            abi = interfaces
-                                .iter()
-                                .map(|iface| iface.source.to_string())
-                                .collect::<Vec<_>>()
-                                .join("\n")
-                                .split("\n")
-                                .map(|line| line.trim().to_string())
-                                .filter(|line| line.starts_with("event"))
-                                .collect();
-                            
-                            // Save to cache in case a future event in the loop is from the same address.
-                            event_abis.insert(address.to_string(), abi.clone());
-                        }
-
-                        // If we get the event ABI from Etherscan...
-                        if &abi.len() > &0 {
-                            // Compare topic[0] to the event abi to find the winner.
-                            let winning_sigs = decode_event_with_abi(topic0, &abi)?;
-                            let winning_sig: Option<&String> = winning_sigs.get(0);                            
-                            
-                            if let Some(winning_sig) = winning_sig {
-                                println!("Contract: {} emits:", &address);
-                                println!("{}", winning_sig);
-                                event_data.sig = Some(winning_sig.clone());
-
-                                // Break the winning signature into its component substrings, and convert to EventArgs.
-                                let mut sig_chunks: Vec<String> = get_sig_chunks(&winning_sig);
-                                sig_chunks.retain(|chunk| !chunk.starts_with("event"));
-                                args = sig_chunks.iter().map(|chunk| chunk.into()).collect();
-                            }    
-                        } 
-
-                        // If the contract wasn't Etherscan verified, try 4byte...
-                        if args.len() == 0 {
-                            let winning_sigs = decode_event_topic(topic0).await?;
-                            let winning_sig: Option<&String> = winning_sigs.get(0);
-
-                            if let Some(winning_sig) = winning_sig {
-                                println!("Contract: {} emits:", &address);
-                                println!("{}", winning_sig);
-                                event_data.sig = Some(winning_sig.clone());
-                            
-                                // Break the 4byte returned signature into its component substrings, and convert to EventArgs.
-                                let sig_chunks: Vec<String> = get_sig_chunks(&winning_sig.trim());
-                                args = sig_chunks[1..]
-                                        .iter()
-                                        .filter(|chunk| chunk.len() > 0)
-                                        .map(|chunk| chunk.into())
-                                        .collect();
-                                
-                                let topics_as_data = topics[1..]
-                                    .iter()
-                                    .map(|topic| topic.as_str().unwrap()[2..].to_string())
-                                    .collect::<Vec<String>>()
-                                    .join("")
-                                    .to_string();
-
-                                data = format!("Ox{}{}", topics_as_data, data[2..].to_string());
-                            }
-                        }
-
-                        // If it wasn't on 4byte either, then just emit the raw topics and data.
-                        if args.len() == 0 {
-                            println!("Contract: {} emits:", &address);
-                            println!("Unknown Event: {}", topic0);
-                            for (idx, topic) in topics[1..].iter().enumerate() {
-                                println!("({}) {}", idx + 1, topic.as_str().unwrap());
-                            }
-                            println!("Data: {}", data);
-                            break; // CAN WE SPECIFY THAT WE WANT TO CONTINUE MAIN LOOP?
-                        };
-
-                        // If we have args, add values and positions to indexed args.
-                        let original_args = args.clone();
-                        let mut args: Vec<EventArg> = args
-                            .iter()
-                            .filter(|arg| arg.indexed)
-                            .zip(&topics[1..])
-                            .zip(1..(topics.len()))
-                            .map(|((arg, topic), count)| {
-                                if let serde_json::Value::String(t) = topic { 
-                                    EventArg {
-                                        name: Some(arg.name.clone().unwrap()),
-                                        type_: arg.type_.clone(),
-                                        indexed: arg.indexed,
-                                        value: Some(convert_value(&arg.type_, t.clone())),
-                                        position: Some(count),
-                                    }
-                                } else {
-                                    EventArg {
-                                        name: Some(arg.name.clone().unwrap()),
-                                        type_: arg.type_.clone(),
-                                        indexed: arg.indexed,
-                                        value: None,
-                                        position: Some(count),
-                                    }
-                                }
-                            })
-                            .collect();
-                        
-                        // Add values to non-indexed args.
-                        let mut non_indexed_args: Vec<EventArg> = original_args
-                            .iter()
-                            .filter(|arg| !arg.indexed)
-                            .zip(split_data_into_args(&data))
-                            .map(|(arg, topic)| {
-                                EventArg {
-                                    name: arg.name.clone(),
-                                    type_: arg.type_.clone(),
-                                    indexed: arg.indexed,
-                                    value: Some(convert_value(&arg.type_, topic.clone())),
-                                    position: None,
-                                }
-                            })
-                            .collect();
-
-                        args.append(&mut non_indexed_args);
-
-                        event_data.args = args;
-
-                        for arg in &event_data.args {
-                            println!("({}) {}: {}", arg.position.clone().unwrap_or(0), arg.name.clone().unwrap_or("Unknown".to_string()), arg.value.clone().unwrap());    
-                        }
-                        
-                        event_return_data.push(event_data);
-                        println!("");
-                    }
-                } else {
-                    eyre::bail!("There's an issue with the tx logs (they don't seem to be an array)!")
-                }
-            } else {
-                // Insert other input options here as they're built.
-                eyre::bail!("Please include a tx hash!")
-            };
+            println!("{}", Cast::new(provider).events(tx_hash, chain.inner, api_key).await?);
         }
         Subcommands::Call { address, sig, args, block, eth } => {
             let config = Config::from(&eth);
@@ -956,122 +772,3 @@ where
     Ok(())
 }
 
-fn translate_event_string_to_hash(event: &str) -> Result<String, eyre::Error> {
-    let sig_chunks: Vec<String> = get_sig_chunks(event);
-
-    let sig = format!("{}({})", 
-        sig_chunks[0].split_whitespace().nth(1).unwrap().to_string(), 
-        sig_chunks[1..].iter().map(|chunk| chunk.split_whitespace().nth(0).unwrap().to_string()).collect::<Vec<String>>().join(","));
-
-    match SimpleCast::keccak(&sig) {
-        Ok(hash) => Ok(hash),
-        Err(e) => Err(e.into()),
-    }
-}
-
-fn get_sig_chunks(event: &str) -> Vec<String> {
-    let mut sig_chunks: Vec<String> = event
-        .split(&[',', '(', ')'][..])
-        .map(|arg| arg.trim().to_string())
-        .collect(); // vec["event Transfer", "address indexed from", "address indexed to", "uint256 value", ";"]
-    
-    sig_chunks.retain(|arg| arg != ";");
-    sig_chunks
-}
-
-struct EventData {
-    sig: Option<String>,
-    address: String,
-    args: Vec<EventArg>,
-}
-
-struct EventArg {
-    name: Option<String>,
-    type_: EventArgType,
-    indexed: bool,
-    value: Option<String>,
-    position: Option<usize>,
-}
-enum EventArgType {
-    Address,
-    Uint256,
-}
-
-fn convert_hex_to_address(hex: &str) -> String {
-    format!("0x{}", hex[hex.len() - 40..].to_string())
-}
-
-fn convert_value(event_type: &EventArgType, value: String) -> String {
-    match event_type {
-        EventArgType::Address => convert_hex_to_address(&value),
-        EventArgType::Uint256 => SimpleCast::to_dec(&value).unwrap().to_string(),
-    }
-}
-
-fn decode_event_with_abi(topic0: &str, abi: &Vec<String>) -> Result<Vec<String>, eyre::Error> {
-    let out = abi
-        .iter()
-        .map(|event_string| (translate_event_string_to_hash(event_string).unwrap(), event_string))
-        .filter(|(hashed_sig, _)| hashed_sig == topic0)
-        .map(|(_, sig)| sig.to_string())
-        .collect::<Vec<String>>();
-
-    Ok(out)
-}
-
-fn split_data_into_args(data: &str) -> Vec<String> {
-    let mut out = Vec::new();
-    let mut open = String::new();
-    for (idx, char) in data.chars().enumerate() {
-        if idx < 2 {
-            continue;
-        } else if idx == 2 || (idx - 1) % 64 != 0 {
-            open.push_str(&char.to_string());
-        } else {
-            open.push_str(&char.to_string());
-            out.push(open);
-            open = String::new();
-        };
-    };
-    // TODO: if there's any leftover in open, error out!
-    out
-}
-
-impl Clone for EventArg {
-    fn clone(&self) -> Self {
-        EventArg {
-            name: self.name.clone(),
-            type_: self.type_.clone(),
-            indexed: self.indexed,
-            value: self.value.clone(),
-            position: self.position.clone(),
-        }
-    }
-}
-
-impl Clone for EventArgType {
-    fn clone(&self) -> Self {
-        match self {
-            EventArgType::Address => EventArgType::Address,
-            EventArgType::Uint256 => EventArgType::Uint256,
-        }
-    }
-}
-
-impl From<&String> for EventArg {
-    fn from(arg: &String) -> Self {
-        let arg_parsed = arg.split_whitespace().collect::<Vec<&str>>();
-        let arg_type = match arg_parsed[0] {
-            "address" => EventArgType::Address,
-            "uint256" => EventArgType::Uint256,
-            _ => panic!("Unsupported event arg type: {}", arg_parsed[0]),
-        };
-        Self {
-            name: if arg_parsed.len() > 1 { Some(arg_parsed.last().unwrap().to_string()) } else { None },
-            type_: arg_type,
-            indexed: arg_parsed.len() == 3,
-            value: None,
-            position: None,
-        }
-    }
-}
